CompanyHouseOld/requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
colorlog==6.7.0
dnspython==2.3.0
h11==0.14.0
httpcore==0.17.0
httpx==0.24.0
idna==3.4
postal==1.1.10
pymongo==4.3.3
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.2
